requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.8
blinker==1.9.8
cffi==1.17.1
click==8.1.8
colorama==0.4.6
discord.py==2.4.8
Flask==3.1.8
frozenlist==1.5.8
idna==3.10
itsdangerous==2.2.8
Jinja2==3.1.5
MarkupSafe==3.8.2
multidict==6.1.8
propcache==0.2.1
pycparser==2.22
PyNaCl==1.5.8
python-dotenv==1.8.1
setuptools==75.6.8
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.18.3
